build: refactor to ESM and single-purpose files
