+++ .gitignore
@@ -1,2 +1,3 @@
 coverage/
+lib/
 node_modules/
