+++ jest.config.json
@@ -6,5 +6,6 @@
       "lines": 100,
       "statements": 100
     }
-  }
+  },
+  "transform": {}
 }
