+++ package.json
@@ -26,16 +26,26 @@
     "@sindresorhus/is": "^3.1.2",
     "char-regex": "^1.0.2",
     "emojilib": "^2.4.0",
-    "skin-tone": "^2.0.0"
+    "skin-tone": "^2.0.0",
+    "tsup": "^6.7.0"
   },
   "devDependencies": {
+    "cross-env": "^7.0.3",
     "jest": "^29.5.0"
   },
   "scripts": {
-    "coverage": "jest --coverage",
+    "build": "tsup ./src/index.js --format esm,cjs --out-dir lib",
+    "coverage": "npm run test -- --coverage",
     "prepublishOnly": "npm run coverage",
-    "test": "jest"
+    "test": "cross-env NODE_OPTIONS=--experimental-vm-modules npx jest"
   },
-  "main": "index.js",
-  "license": "MIT"
+  "exports": {
+    ".": {
+      "import": "./lib/index.js",
+      "require": "./lib/index.cjs"
+    }
+  },
+  "main": "./lib/index.js",
+  "license": "MIT",
+  "type": "module"
 }
