+++ src/strip.js
@@ -1,0 +1,4 @@
+import { replace } from './replace.js'
+
+export const strip = (input, { preserveSpaces } = {}) =>
+  replace(input, '', { preserveSpaces })
